Removed parasitics and tuned
After a meeting with Dr. Lemmon, we learned that Resistors and Inductors both experience frequency-dependent behavior as well. We tuned this with the assumption that they were static, but no longer.  I will begin working on getting those frequency dependencies accounted for, with the help of Dr. Lemmon and Dr. Courtay